+++ kernel-rs/src/plic.rs
@@ -8,8 +8,8 @@ use crate::{
 
 pub unsafe fn plicinit() {
     // set desired IRQ priorities non-zero (otherwise disabled).
-    *((PLIC + (UART0_IRQ * 4) as i64) as *mut u32) = 1;
-    *((PLIC + (VIRTIO0_IRQ * 4) as i64) as *mut u32) = 1;
+    *((PLIC as i64 + (UART0_IRQ * 4) as i64) as *mut u32) = 1;
+    *((PLIC as i64 + (VIRTIO0_IRQ * 4) as i64) as *mut u32) = 1;
 }
 
 pub unsafe fn plicinithart() {
